+++ scripts/macros/2024/classFeatures/monk/heightenedFocus.js
@@ -1,0 +1,15 @@
+export let heightenedFocus = {
+    name: 'Heightened Focus',
+    version: '1.3.141',
+    rules: 'modern',
+    config: [
+        {
+            value: 'attacks',
+            label: 'CHRISPREMADES.Config.Attacks',
+            type: 'number',
+            default: 1,
+            category: 'homebrew',
+            homebrew: true
+        }
+    ]
+};